refactor(msa): review-driven cleanup of the MSA PR surface
File organization:
- Rename the prefill path to its canonical name: sparse_attention.py ->
  sparse_prefill.py, sparse_fwd_union_sm12x.py -> sparse_prefill_sm12x.py
  (class SparsePrefillSm12x), test_sparse_union.py -> test_sparse_prefill.py
- Merge single-consumer aux kernels into their consumers, matching the
  two-classes-per-file shape of proxy_score_sm12x.py: BuildUnionMetaSm12x ->
  sparse_prefill_sm12x.py, SparseCombineSm12x -> sparse_decode_sm12x.py
- New _common.py for helpers shared by the op wrappers; combine helpers move
  from the prefill wrapper into sparse_decode.py; the device union-metadata
  builder folds into sparse_prefill.py (its only caller)
- Move test-only torch oracles out of the package into tests/msa_ops
  (union-metadata host builder, _combine_partials_torch)
- cute_dsl/__init__.py follows the internal-package convention (no exports)

Dead code:
- Delete build_k2q_csr_sm12x.py (leftover of the removed kv-major path)
- Drop the unused max_k_tiles parameter from both top-k kernels
- Drop the inherited _MAX_K_TILES=12288 cap: the multi-stage radix select is
  capacity-free by construction (an over-full threshold bin falls through to
  the next stage), verified adversarially far beyond the old bound; new
  regression test pins 16384 blocks with a value-based check

Comments and docstrings:
- Class/module docstrings lead with the design rationale instead of grid or
  work-item mechanics; parameter contracts live in the public wrapper
  docstrings
- Remove reference-kernel/port-source mentions, tuning-session shorthand,
  benchmark numbers, and narration of standard idioms (smem-reuse barriers,
  double-buffering); keep footguns, invariants, and derivations
- License headers added to test_sparse_prefill.py and test_cuda_graph.py;
  header format matches the house <license>\n<summary> docstring convention

Trace templates and regenerated fi_trace_out JSONs follow the doc changes.
Full msa_ops suite passes (98); trace tests pass; lint clean.

AI-assisted (Claude Code), with every change directed and reviewed manually.
