+++ flashinfer/msa_ops/cute_dsl/__init__.py
@@ -1,5 +1,21 @@
-from .sparse_combine_sm12x import SparseCombineSm12x
+# Copyright (c) 2026 by FlashInfer team.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""
+MSA CuTe-DSL Kernels (internal)
+===============================
 
-__all__ = [
-    "SparseCombineSm12x",
-]
+Internal module containing the SM120/SM121 MSA kernel implementations.
+Import from ``flashinfer.msa_ops`` for the public API; wrappers import
+kernel classes from the submodules directly (and lazily).
+"""
